requirements.txt
absl-py==0.4.1
astor==0.7.1
certifi==2016.2.28
cycler==0.10.0
gast==0.2.0
grpcio==1.15.0
h5py==2.9.0
jieba==0.39
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==2.2.3
numpy==1.14.5
pandas==0.23.4
protobuf==3.6.1
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
scikit-learn==0.19.2
scipy==1.1.0
six==1.11.0
sklearn==0.0
tensorboard==1.10.0
tensorflow-gpu==1.10.0
termcolor==1.1.0
Werkzeug==0.14.1
wincertstore==0.2
xlwt==1.3.0
